backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib
python-dotenv
pydantic-settings
python-multipart
